pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ocr-agent"
version = "0.1.0"
description = "OCR agent for PDFs with scanned images - supports OpenAI, Claude, Ollama via LiteLLM"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "Toystack" }]
keywords = ["ocr", "pdf", "document", "llm", "litellm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "python-dotenv>=1.0",
    "pypdf>=4.0",
    "pdf2image>=1.16",
    "PyMuPDF>=1.24",
    "Pillow>=10.0",
    "litellm>=1.30",
    "langchain-text-splitters>=0.2",
    "tiktoken>=0.5",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "python-multipart>=0.0.9",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]
postgres = ["pgvector>=0.2", "asyncpg>=0.29", "psycopg[binary]>=3.0"]
qdrant = ["qdrant-client>=1.7"]
tesseract = ["pytesseract>=0.3.10"]
all = [
    "ocr-agent[postgres,qdrant,tesseract]",
]

[project.scripts]
ocr-agent = "ocr_agent.cli:main"

[project.urls]
Documentation = "https://github.com/toystack/ocr-agent"
Repository = "https://github.com/toystack/ocr-agent"

[tool.setuptools.packages.find]
where = ["src"]
